Enable `NoHashHasher` for custom types.
The current implementation of this crate is overly restrictive and
prevents using `NoHashHasher` with `HashMap`s whose domain type is
not one of the built-in integers. The `Hasher` impls of `NoHashHasher`
for those integers enforce a 1:1 correspondence between the specific
integer parameter type and the `Hasher::write_*` method that is invoked
for this type, e.g. `NoHashHasher<u32>` only allows a single call to
`Hasher::write_u32`. Not just double-invocations are considered an
error, but also a single hypothetical call to `Hasher::write_u64`
instead of `Hasher::write_u32`.

In here we argue that this restriction is unnecesary. The important
invariant `NoHashHasher` needs to enforce is that *one* of the `Hasher`
methods is invoked exactly once (with the exception of `write` which
accepts a `&[u8]` as parameter), but it is not important which one
specifically as this is up to the impl of `Hash` to decide. Once we
allow calling any method we can replace the multitude of `NoHashHasher`
impls for each of the integer types with a generic one and we are almost
done.

The only remaining aspect is that we do not want to accept just any
type for `NoHashHasher` as the type system does not capture the
constraint of invoking a `Hasher` method exactly once. We therefore
add a marker trait `IsEnabled` and a corresponding bound to the
`Hasher` impl for `NoHashHasher` to ensure that only types which
have opted in are considered.

With these changes in place, `HashMap`s and `HashSet`s with custom
types are now easily defined. See the doc comment of `IsEnabled`
for an example.